backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.1
email-validator==2.1.0
httpx==0.25.1
requests==2.31.0
redis==5.0.1
celery==5.3.4
google-maps-services==1.12.0
openai==1.3.6
sqlalchemy-utils==0.41.1
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
isort==5.13.2
